refactor broser-exteionsion/source/react-lowcode
