+++ packages/browser-extension/source/react-lowcode/functions/rename.ts
@@ -1,0 +1,13 @@
+import { Project } from "ts-morph"
+
+export const renameFunctionWithMorph = (
+  project: Project,
+  code: string,
+  functionName: string,
+  newFunctionName: string
+) => {
+  const newProject = project.createSourceFile(`${newFunctionName}.tsx`, code)
+  const mainFunction = newProject.getFunction(functionName)
+  mainFunction?.rename(newFunctionName)
+  return newProject
+}
