+++ packages/browser-extension/source/react-lowcode/io/dir.ts
@@ -1,6 +1,6 @@
     
 export interface CodeDir {
-    readDirectory(path: string, /*extensions?: readonly string[], exclude?: readonly string[], include?: readonly string[], depth?: number*/): Promise<string[] | undefined>
+    readDirectory(path: string, extensions?: readonly string[], exclude?: readonly string[], include?: readonly string[], depth?: number): Promise<string[] | undefined>
 
     // getCurrentDirectory: () => "/",
     //
