--- packages/browser-extension/source/tsx/mapRecognition.test.ts
@@ -1,9 +1,0 @@
-import { functionalComponent } from "../util/code"
-import { astFindStart } from "./ast"
-import { isInsideMapPatternFunction } from "./mapRecognition"
-
-test("should ", () => {
-  const node = astFindStart(functionalComponent, 196)
-
-  expect(isInsideMapPatternFunction(node!)).toBeTruthy()
-})
